1.0.4-RELEASE
The displayName associated with the targetSiteCode is null and it
appears that there is a missing check for codeNameAvailable in the
AbstractVocabFunction class on the line where the NullPointerException
is pointed to in the lambda function mentioned in the stack trace. I've
added the check to my local working copy of the CRV code (see attached
patch) (note: I have version 1.0.3-SNAPSHOT as my working copy, which is
an older version than what you probably have/is deployed at NIST). By
adding that check (codeNameAvailable is true if codeName != null),
running the sample file should not display the stack trace.